Updated the build/master.project Makefiles so that the machine-id is generated and used during the entire build process.
This changes the boot-command used for building the box so that we can seed systemd
with the generated machine-id. As CoreOS does not have a forgiving timeout, we have
to literally mash keys and then hit backspace in order to catch the boot loader in
order to type in our id at the commandline. This machine-id is also fed into the
ignition file in order to seed the box with it on first boot.

The intention of these changes are so that systemd's dhcp client will retain the
same address during the entire build process. All of this is to work around issue
hashicorp/packer#8528 and is pretty much because Packer is unable to re-determine
the address of the guest if it changes whilst in the middle of building. It turns
out that the latest version of CoreOS resets the machine-id after the install which
results in this specific issue.